2024-03-19 view like comments

+++ server/api/addLike.php
@@ -7,22 +7,17 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         $uid = $_SESSION['user']['uid'];
         $postId = $_POST['postId'];
 
-        $selectAll = "SELECT * FROM likes WHERE post_id = ?";
-        $stmt = $connection->prepare($selectAll);
-        $stmt->bind_param("i", $postId);
-        $stmt->execute();
-        $result = $stmt->get_result();
+        $selectAll = "SELECT * FROM likes WHERE post_id = '$postId' AND liked_by = '$uid'";
+        $result = mysqli_query($connection, $selectAll);
+        
 
-        if ($result->num_rows == 0) {
-            $insertLike = "INSERT INTO `likes`(`post_id`, `liked_by`, `created_date_time`) VALUES (?, ?, now())";
-            $stmt = $connection->prepare($insertLike);
-            $stmt->bind_param("ii", $postId, $uid);
-            $stmt->execute();
+        if (mysqli_num_rows($result) == 0) {
+            $insertLike = "INSERT INTO `likes`(`post_id`, `liked_by`, `created_date_time`) VALUES ('$postId', '$uid', now())";
+            $addLikeStatus = mysqli_query($connection, $insertLike);
+            
         } else {
-            $deleteLike = 'DELETE FROM `likes` WHERE `liked_by` = ? AND `post_id` = ?';
-            $stmt = $connection->prepare($deleteLike);
-            $stmt->bind_param("ii", $uid, $postId);
-            $stmt->execute();
+            $deleteLike = "DELETE FROM `likes` WHERE `liked_by` = '$uid' AND `post_id` = '$postId'";
+            $deleteLikeStatus = mysqli_query($connection, $deleteLike);
         }
     } else {
         echo "User not logged in.";
